Improve modularization (#69)
Closes #2.
