+++ piplicenses/errors.py
@@ -2,7 +2,8 @@
 # SPDX-FileCopyrightText: Copyright (c) 2018 raimon
 # SPDX-FileCopyrightText: Copyright (c) 2025 stefan6419846
 
-if __name__ == "__main__":  # pragma: no cover
-    from piplicenses.cli import main
 
-    main()
+class PipLicensesWarning(UserWarning):
+    """
+    Base class for warnings emitted by the pip-licenses-cli package.
+    """
